requirements.txt
numpy
pandas
matplotlib
scikit-learn
pytorch
torchvision
monai
nibabel
SimpleITK
streamlit
dash
plotly
